feat: Rotate the captured images to correct angle
Images of the boards were captured with a camera on a selfie stick
mouted on a table inbetween books to hold the selfie stick.
Not the perfect setup. Naturally pictures taken weren't perfect either.
For a start, the pictures had to be rotated to the correct angle.
Some had to be rotated 90 degree and some had to be rotated 270 degree.