feat: Align captured images
Images of the boards were captured with a camera on a selfie stick
mouted on a table inbetween books to hold the selfie stick. Not the
perfect setup. Naturally pictures taken weren't perfect either.

One problem is because of small movement in the camera, the location
of the board moves a little in the images. Also, camera turns a little
and we end up capturing the board from different angles.

There are computer vision techniques available to correct these
effects. We'll use OpenCV's ECC image alignment feature to align the
captured images to create images as if the camera never moved while
taking all these images.

This is important because we'll extract the edges from these images and
place them on blank board to create any board configuration we want.
If the images are not aligned, we will end up with blocks misplaced on
board.