feat: Remove contours with too little area
masks.py does math operations on images and computes a mask.
However because of small movements in camera, there will be tiny
white regions all over the map. These are not areas where the images
truly differ. These need to be removed to be able to use these images
to generate boards.

This module computes contours in the mask and removes contours with
area less than a certain threshold.