feat: Create board and mask numpy files
During training, it is convinient to not have to read from JPEG files
and decode every single image. We'll convert the images into numpy
arrays and persist in files. We'll read these files directly during
training.

We'll resize the images to 1012x759 to strike a balance between too
big images that takes too much memory and too small images that could
loose details necessary to recognize successfully.

We'll store only the part of the image that actually contains the graph