feat: Create data required to create .rec file
Between speed of training and speed at which we can generate synthetic
data, the later is the bottleneck. To make sure time on GPU instance is
spent wisely, we'll generate the training images and store them into
.rec files before we start training. Training process can then read
directly from .rec file. This way we can run training as fast as the GPU
instance allows.